requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
plotly
matplotlib
seaborn
python-dotenv
fpdf
numpy